KYC

Description

This module contains the KYC data type, as well as examples of user data and a contract that validates this data.

Documentation

data KYCData n k r context

KYC data type

Constructors

KYCData

Fields

kycType :: ByteString k context

KYC Type

kycID :: UInt k r context

Id of KYCData object

kycHash :: UInt k r context

Hash of kycValue field

kycValue :: ByteString n context

ByteString which contains user data in serialized form

data User r context

User data type In contracts it can be deserialized from kycValue

Constructors

User

Fields

userAge :: UInt 8 r context

User age

userCountry :: ByteString 10 context

User country information

kycExample :: forall n k r rsc context. (Symbolic context, KnownNat rsc, KnownRegisterSize r, KnownNat (NumberOfRegisters (BaseField context) 8 r), KnownNat (Ceil (GetRegisterSize (BaseField context) 8 r) OrdWord), KnownNat (NumberOfRegisters (BaseField context) k r), 8 <= n, 10 <= n, KnownNat (n - 10)) => KYCData n k r context -> UInt k r context -> Bool context

Example of KYC contract. Includes

hash validation

user age verification (> 18)

user country is not in the list of restrictions

userA :: forall r context. (Symbolic context, KnownRegisterSize r) => User r context

Example of user data Example user data. An example serialized representation of this data can be seen in example-json/kyc-data.json